requirements.txt
alabaster==0.7.11
asn1crypto==0.24.0
Babel==2.6.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3
Django==2.0.7
django-bootstrap3==10.0.1
django-crispy-forms==1.7.2
django-filter==2.0.0
django-oauth-toolkit==1.2.0
django-rest-swagger==2.2.0
django-xadmin==0.5.0
djangorestframework==3.8.2
docutils==0.14
idna==2.7
imagesize==1.0.0
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.13
oauthlib==2.1.0
openapi-codec==1.3.2
packaging==17.1
pycparser==2.18
Pygments==2.2.0
PyMySQL==0.9.2
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.6
sphinx-rtd-theme==0.4.1
sphinxcontrib-websupport==1.1.0
uritemplate==3.0.0
urllib3==1.23
